os: remove conversion warnings/errors from virtio
The read_config and write_config functions in the generic virtio
headers used by all drivers lead to compiler warnings resp. errors
if effective-c++ switch is enabled. Moreover, the functions require
to define the access width as parameter. We can better turn them
into template functions using the value type to read resp. write to
derive the access width.

Ref genodelabs/genode#4344

+++ repos/os/include/virtio/queue.h
@@ -413,7 +413,7 @@ class Virtio::Queue
 			static_assert(!TRAITS::device_write_only);
 			static_assert(TRAITS::has_data_payload);
 
-			int const req_desc_count = 1 + (sizeof(header) + data_size) / _buffers.buffer_size();
+			size_t const req_desc_count = 1UL + (sizeof(header) + data_size) / _buffers.buffer_size();
 			if (req_desc_count > _descriptors.available_capacity())
 				return false;
